fix: denomination step values
Investigating the issue where a user cannot select denominations of denominations i.e 0.1 milli or 0.01 whole.

The problem was coming from the b-input tag and it did not have any value for step. Without a value the range is whole numbers.
Added a step with 3 decimal places as that is what is being used in formatBalance.
This in turn removes the formatting issue.
Fixes vue-polkadot/apps#63
After verfiying the fix makes the problem go away I went to search for other components which use it and found only 1 in the Amount component.